+++ packages/css-color-parser/src/gamut-mapping/srgb.ts
@@ -1,4 +1,5 @@
-import { clip, XYZ_D50_to_sRGB, OKLCH_to_OKLab, XYZ_to_lin_sRGB, XYZ_D50_to_OKLCH, XYZ_to_OKLab, OKLab_to_OKLCH, gam_sRGB, lin_sRGB, OKLab_to_XYZ, mapGamut, lin_sRGB_to_XYZ, inGamut, Color } from '@csstools/color-helpers';
+import { mapGamutRayTrace } from '@csstools/color-helpers';
+import { clip, XYZ_D50_to_sRGB, OKLCH_to_OKLab, XYZ_to_lin_sRGB, XYZ_D50_to_OKLCH, XYZ_to_OKLab, OKLab_to_OKLCH, gam_sRGB, OKLab_to_XYZ, lin_sRGB_to_XYZ, inGamut, Color } from '@csstools/color-helpers';
 
 export function XYZ_D50_to_sRGB_Gamut(color: Color): Color {
 	const srgb = XYZ_D50_to_sRGB(color);
@@ -16,15 +17,17 @@ export function XYZ_D50_to_sRGB_Gamut(color: Color): Color {
 		oklch = [1, 0, 0] as Color;
 	}
 
-	return mapGamut(oklch, (x: Color) => {
-		x = OKLCH_to_OKLab(x);
-		x = OKLab_to_XYZ(x);
-		x = XYZ_to_lin_sRGB(x);
-		return gam_sRGB(x);
-	}, (x: Color) => {
-		x = lin_sRGB(x);
-		x = lin_sRGB_to_XYZ(x);
-		x = XYZ_to_OKLab(x);
-		return OKLab_to_OKLCH(x);
-	});
+	return gam_sRGB(mapGamutRayTrace(oklch, oklch_to_lin_srgb, lin_srgb_to_oklch));
+}
+
+function oklch_to_lin_srgb(x: Color): Color {
+	x = OKLCH_to_OKLab(x);
+	x = OKLab_to_XYZ(x);
+	return XYZ_to_lin_sRGB(x);
+}
+
+function lin_srgb_to_oklch(x: Color): Color {
+	x = lin_sRGB_to_XYZ(x);
+	x = XYZ_to_OKLab(x);
+	return OKLab_to_OKLCH(x);
 }